Revert "VertexShaderStage"
This reverts commit 177d8c521865393b168623918c3520a303a78046.
